refactor: consolidate URL constants into metadata module
Moved PRODUCTION_URL and BASE_URL from constants/links to constants/metadata for better organization. Inlined specific links like waitlistFormLink and fds7Link directly in components to reduce dependency on the links module. This simplifies imports and centralizes metadata-related constants.
